+++ backend/src/main/java/com/github/ingbeck/backend/model/appuser/AppUserCreateDto.java
@@ -1,4 +1,4 @@
-package com.github.ingbeck.backend.model;
+package com.github.ingbeck.backend.model.appuser;
 
 public record AppUserCreateDto(
         String birthdate,
