+++ backend/src/main/java/com/github/ingbeck/backend/model/openfoodfactsresponse/Nutriments.java
@@ -1,4 +1,4 @@
-package com.github.ingbeck.backend.model;
+package com.github.ingbeck.backend.model.openfoodfactsresponse;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import lombok.AllArgsConstructor;
