+++ backend/src/main/java/com/github/ingbeck/backend/service/UserService.java
@@ -1,8 +1,8 @@
 package com.github.ingbeck.backend.service;
 
-import com.github.ingbeck.backend.model.AppUser;
-import com.github.ingbeck.backend.model.AppUserCreateDto;
-import com.github.ingbeck.backend.model.AppUserGender;
+import com.github.ingbeck.backend.model.appuser.AppUser;
+import com.github.ingbeck.backend.model.appuser.AppUserCreateDto;
+import com.github.ingbeck.backend.model.appuser.AppUserGender;
 import com.github.ingbeck.backend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
